+++ apps/io-services-cms-webapp/src/main.ts
@@ -411,3 +411,4 @@ export const createRequestServicesLifecycleIngestionRetryEntryPoint =
   createRequestServicesLifecycleIngestionRetryHandler(
     serviceLifecycleEventHubProducer,
   );
+