+++ apps/backoffice/CHANGELOG.md
@@ -1,5 +1,17 @@
 # io-services-cms-backoffice
 
+## 1.15.4
+
+### Patch Changes
+
+- 0fbc0b1: fix manage subscription creation
+
+## 1.15.3
+
+### Patch Changes
+
+- a05f4fe: retrive service list: resolve group name
+
 ## 1.15.2
 
 ### Patch Changes